dev-requirements.txt
# task runner
invoke>=0.13.0

# testing
pytest==2.8.0
tox>=1.5.0
mock

# packaging
wheel
twine

# Syntax checking
flake8==2.4.1

# Soft requirements
flask-sqlalchemy
marshmallow-sqlalchemy>=0.13.0
